azure-neuromodels-project/src/functions/analyze_sentiment/requirements.txt
azure-ai-textanalytics
azure-cosmos
